code refactor as per pr
